+++ scraper/package.json
@@ -1,0 +1,6 @@
+{
+  "dependencies": {
+    "jsdom": "^22.0.0",
+    "mysql": "^2.18.1"
+  }
+}
